Added big food

+++ Assets/Code/WebModule/StickierWeb.cs
@@ -2,9 +2,9 @@ namespace ThroughAThousandEyes.WebModule
 {
     public class StickierWeb : Upgrade
     {
-        public double ExtraTime => Level * _extraTimePerLevel;
+        public float ExtraTime => Level * _extraTimePerLevel;
 
-        private double _extraTimePerLevel;
+        private float _extraTimePerLevel;
         
         public StickierWeb(WebModuleData.UpgradeData data, WebModuleRoot root) : base(data, root)
         {
